+++ src/domain/transaction/TransactionCreator.java
@@ -2,7 +2,8 @@ package domain.transaction;
 
 import domain.Transaction;
 import domain.TransactionType;
+import java.math.BigDecimal;
 
 public interface TransactionCreator {
-    Transaction create(double amount, TransactionType type, String accountNumber);
+    Transaction create(BigDecimal amount, TransactionType type, String accountNumber);
 }
